Move Ramm/helpers to utils/internalPrice
